Support fractional-second intervals on recurring timers
A fractional-second timeout was already supported for the initial timeout of a /TIMER, but
if the timer was recurring then the interval was rounded to a whole number of seconds for
the subsequent timeouts.

Change the type of interval from long to double to fix this.

+++ source/timer.c
@@ -421,8 +421,8 @@ char *function_timer(char *n, char *args)
 	if (time_left < 0)
 		time_left = 0.0;
 
-	return m_sprintf("%s %d %d %ld %d %0.3f %s%s", t->ref, t->server,
-		t->window, (long)t->interval, t->events, time_left, 
+	return m_sprintf("%s %d %d %.16g %d %0.3f %s%s", t->ref, t->server,
+		t->window, t->interval, t->events, time_left, 
 		t->callback ? "(internal callback) " : t->command,
 		t->whom ? t->whom : "");
 }
